Component PersonListView moved into folder app/components.
